requirements.txt
notion-client==2.2.1
supabase==2.3.4
python-dotenv==1.0.1
schedule==1.2.1 